oauth2-authenticator/pyproject.toml
[tool.poetry]
name = "oauth2-authenticator"
version = "0.1.0"
description = "arXiv user admin API"
authors = ["arXiv <developer@arxiv.org>"]
license = "BSD-3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
arxiv-base = {git = "https://github.com/arXiv/arxiv-base.git", rev = "cc0597aa421fe355d92f9ad352a43292d78db638"}
pymysql = "^1.1.0"
ruamel-yaml = "^0.18.6"
python-json-logger = "^2.0.7"
python-dateutil = "^2.8.2"
pydantic-sqlalchemy-2 = "^0.0.2"
pyjwt = "^2.9.0"
requests = "^2.32.3"
jwcrypto = "^1.5.6"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
